% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_observed.R
\name{average_observed}
\alias{average_observed}
\title{Average Observed}
\usage{
average_observed(
  x_name,
  data,
  y,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  ...
)
}
\arguments{
\item{x_name}{Column name of the variable shown on the x axis.}

\item{data}{Matrix-like.}

\item{y}{Numeric vector with observed values of the response.
Can also be a column name in \code{data}. Omitted if \code{NULL} (default).}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{x_name}. Only relevant for numeric x.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{x_name}. Only relevant for numeric x.}

\item{discrete_m}{Numeric x with up to this number of unique values
should be treated as factors. The default is 2. Vectorized over \code{x_name}.}

\item{winsorize_low}{Small values of numeric x are capped at this quantile.
Set to 0 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{x_name}.}

\item{winsorize_high}{High values of numeric x are capped at this quantile.
Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{x_name}.}

\item{...}{Currently unused.}
}
\value{
If \code{x_name} has length 1, an object of class "marginal" containing these elements:
\itemize{
\item \code{data}: data.frame containing statistics and plot positions of values and bars.
\item \code{num}: Indicator whether x is numeric.
\item \code{x_name}: Same as input \code{x_name}.
If \code{x_name} has length > 1, an object of class "multimarginal", which is a named
list of "marginal" objects.
}
}
\description{
Calculates average observed \code{y} values over one or multiple x variables specified
via \code{x_names}. \code{y} can either be a numeric vector or a column name in \code{data}.
This function is a convenience wrapper over \code{\link[=marginal]{marginal()}}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
M <- average_observed(
  x_name = "Sepal.Width", y = "Sepal.Length", data = iris, breaks = 5
)
M
M |> plot()

xvars <- c("Sepal.Width", "Species")
M <- average_observed(xvars, data = iris, y = "Sepal.Length")
M$Species
plot(M, rotate_x = 45)
}
\seealso{
\code{\link[=marginal]{marginal()}}
}
