% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch.R
\name{.ale}
\alias{.ale}
\title{Barebone Accumulated Local Effects}
\usage{
.ale(
  object,
  v,
  X,
  breaks,
  right = TRUE,
  pred_fun = stats::predict,
  trafo = NULL,
  which_pred = NULL,
  max_bin_size = 50L,
  w = NULL,
  g = NULL,
  ...
)
}
\arguments{
\item{object}{Fitted model.}

\item{v}{Variable name in \code{data} to calculate ALE.}

\item{X}{Matrix or data.frame.}

\item{breaks}{Breaks for ALE calculation.}

\item{right}{Should bins specified via \code{breaks} be right-closed?
The default is \code{TRUE}.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{trafo}{How should predictions be transformed?
A function or \code{NULL} (default). Examples are \code{log} (to switch to link scale)
or \code{exp} (to switch from link scale to the original scale).}

\item{which_pred}{If the predictions are multivariate: which column to pick
(integer or column name). By default \code{NULL} (picks last column).}

\item{max_bin_size}{Maximal number of observations used per bin. If there are more
observations in a bin, \code{n_bin} indices are randomly sampled. The default is 50.}

\item{w}{Optional vector with case weights.}

\item{g}{For internal use. The result of \code{factor(findInterval(...))}.
By default \code{NULL}.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} or (typically) \code{prob = TRUE} in classification models.}
}
\value{
Vector of ALE values in the same order as \code{breaks[-length(breaks)]}.
}
\description{
Very fast ALE calculations. Since the function does not do any
input checks, it is mainly meant for internal use. Still, for developers, it can
be handy.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
v <- "Sepal.Width"
.ale(fit, v, X = iris, breaks = seq(2, 4, length.out = 5))
}
\seealso{
\code{\link[=partial_dependence]{partial_dependence()}}
}
