% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{ep_importance}
\alias{ep_importance}
\title{Variable Importance}
\usage{
ep_importance(x, by = NULL)
}
\arguments{
\item{x}{Object of class "marginal".}

\item{by}{The statistic used to calculate the variance for.
One of 'pd', 'pred_mean', 'y_mean', 'resid_mean', or 'ale' (if available).
The default is \code{NULL}, which picks the first available statistic from above list.}
}
\description{
Extracts from a "marginal" object a simple variable importance measure, namely
the (bin size weighted) variance of the partial dependence values, or of any other
calculated statistic (e.g., "pred_mean" or "y_mean"). It can be used via
\code{update.marginal(, sort_by = "pd")} to sort the variables in decreasing importance.
Note that this measure captures only the main effect strength.
If the importance is calculated with respect to "pd", it is closely related
to the suggestion of Greenwell et al. (2018).
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
M <- marginal(fit, v = colnames(iris)[-1], data = iris)
ep_importance(M)
}
\references{
Greenwell, Brandon M., Bradley C. Boehmke, and Andrew J. McCarthy. 2018.
\emph{A Simple and Effective Model-Based Variable Importance Measure.}
arXiv preprint. \url{https://arxiv.org/abs/1805.04755}.
}
\seealso{
\code{\link[=update.marginal]{update.marginal()}}
}
