% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/marginalplot.R
\name{marginal}
\alias{marginal}
\alias{marginal.default}
\alias{marginal.ranger}
\alias{marginal.explainer}
\title{Marginal Statistics}
\usage{
marginal(object, ...)

\method{marginal}{default}(
  object,
  x_name,
  data,
  y = NULL,
  pred = NULL,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

\method{marginal}{ranger}()

\method{marginal}{explainer}()
}
\arguments{
\item{object}{Fitted model.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} model.}

\item{x_name}{Column name of the stratification variable shown on the x axis.}

\item{data}{Matrix-like.}

\item{y}{Numeric vector with observed values of the response.
Can also be a column name in \code{data}. Omitted if \code{NULL} (default).}

\item{pred}{Numeric vector with predictions. If \code{NULL}, it is calculated as
\code{pred_fun(object, data, ...)}. Used to save time if the function is to be
called multiple times.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{x_name}.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{x_name}.}

\item{discrete_m}{Numeric variable with up to this number of unique values
should not be binned. The default is 2. Set to \code{Inf} to avoid any binning.
Vectorized over \code{x_name}.}

\item{calc_pred}{Should predictions \code{pred} be calculated? Default is \code{TRUE}.}

\item{pd_n}{Size of the dataset used for calculation of partial dependence.
The default is 500. Set to 0 (or pass \code{pred_fun = NULL}) to omit calculation
of partial dependence.}

\item{wins_low}{High values of \code{x} are capped at the \code{wins_high} quantile.
Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile. Vectorized over \code{x_name}.}
}
\value{
An object of class "marginal" containing these elements:
\itemize{
\item \code{data}: data.frame containing the partial dependencies.
\item \code{v}: Same as input \code{v}.
\item \code{K}: Number of columns of prediction matrix.
\item \code{pred_names}: Column names of prediction matrix.
\item \code{by_name}: Column name of grouping variable (or \code{NULL}).
}
}
\description{
Calculates
\itemize{
\item average observed,
\item average predicted,
\item partial dependence, and
\item counts/weights
over a (binned) feature v, possibly weighted.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{marginal(default)}: Default method.

\item \code{marginal(ranger)}: Method for "ranger" models.

\item \code{marginal(explainer)}: Method for DALEX "explainer".

}}
\examples{
library(ranger)

fit <- ranger(Sepal.Length ~ ., data = iris)
xvars <- c("Sepal.Width", "Petal.Width", "Petal.Length", "Species")
M <- marginal.default(fit, x_name = xvars, data = iris, y = "Sepal.Length")
M$Petal.Width
M |> plot()
M |> plot(backend = "plotly")
}
\references{
Friedman, Jerome H. \emph{"Greedy Function Approximation: A Gradient Boosting Machine."}
Annals of Statistics 29, no. 5 (2001): 1189-1232.
}
