% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\alias{partial_dependence.default}
\alias{partial_dependence.ranger}
\alias{partial_dependence.explainer}
\title{Partial Dependence}
\usage{
partial_dependence(object, ...)

\method{partial_dependence}{default}(
  object,
  x_name,
  data,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{ranger}(
  object,
  x_name,
  data,
  pred_fun = NULL,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{explainer}(
  object,
  x_name,
  data = object[["data"]],
  pred_fun = object[["predict_function"]],
  w = object[["weights"]],
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)
}
\arguments{
\item{object}{Fitted model.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} model.}

\item{x_name}{Column name of the stratification variable shown on the x axis.}

\item{data}{Matrix-like.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{x_name}.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{x_name}.}

\item{discrete_m}{Numeric variable with up to this number of unique values
should not be binned. The default is 2. Set to \code{Inf} to avoid any binning.
Vectorized over \code{x_name}.}

\item{pd_n}{Size of the dataset used for calculation of partial dependence.
The default is 500. Set to 0 (or pass \code{pred_fun = NULL}) to omit calculation
of partial dependence.}
}
\description{
Calculates partial dependence of one or more variables (\code{x_name}) in
a dataset \code{data}. This function is a convenience wrapper over \code{\link[=marginal]{marginal()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{partial_dependence(default)}: Default method.

\item \code{partial_dependence(ranger)}: Default method.

\item \code{partial_dependence(explainer)}: Default method.

}}
\examples{
library(ranger)

fit <- ranger(Sepal.Length ~ ., data = iris)
M <- partial_dependence(fit, x_name = "Species", data = iris)
M
M |> plot()
}
