% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.marginal}
\alias{plot.marginal}
\title{Plots "marginal" Object}
\usage{
\method{plot}{marginal}(
  x,
  ncols = 2L,
  share_y = FALSE,
  sort = FALSE,
  ylim = NULL,
  scale_exposure = 1,
  line_colors = getOption("marginalplot.line_colors"),
  fill = getOption("marginalplot.fill"),
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)
}
\arguments{
\item{x}{An object of class "marginal".}

\item{ncols}{Number of columns in the plot layout.
Only if \code{length(x) > 1} (multiple plots). With "plotly" subplots, only nrow can
be set. Therefore, the calculation is approximately possible only.}

\item{share_y}{Should y axis be shared across all subplots?
No effect if \code{ylim} is passed. Only if \code{length(x) > 1} (multiple plots).}

\item{sort}{Should plots be sorted in decreasing order of importance? Importance is
measured by the exposure weighted variance of the most relevant available statistic
(pd > pred > obs). Only if \code{length(x) > 1} (multiple plots).}

\item{ylim}{Manual y axis range.}

\item{scale_exposure}{Vertical scaling of the exposure bars (between 0 and 1).
The default is 1. Set to 0 for no bars. With "plotly", values between 0 and 1 ar
currently not possible.}

\item{line_colors}{Named vector of line colors. By default, a color blind
palette from "ggthemes" is used, equaling to
\code{c(obs = "#E69F00", pred = "#009E73", pd = "#56B4E9")}.
To change globally, set \code{options(marginalplot.line_colors = "named color vector")}.
Can be used to remove certain lines in the plot.}

\item{fill}{Fill color of bars. The default equals "lightgrey".
To change globally, set \code{options(marginalplot.fill = "new color")}.}

\item{wrap_x}{Should categorical xaxis labels be wrapped after this length?
The default is 10. Set to 0 for no wrapping. Vectorized over \code{x}.
Only for "ggplot2" backend.}

\item{rotate_x}{Should categorical xaxis labels be rotated by this angle?
The default is 0 (no rotation). Vectorized over \code{x}. Only for "ggplot2" backend.}

\item{backend}{Plot backend, either "ggplot2" (default) or "plotly".
To change globally, set \code{options(marginalplot.backend = "plotly")}.}

\item{...}{Currently unused.}
}
\value{
If \code{length(x) == 1} (single plot), an object of class  "ggplot" or "plotly".
Otherwise, an object of class "patchwork" or a "plotly" subplot.
}
\description{
Plots all statistics using a color blind palette from "ggthemes".
You can switch to the interactive interface by setting \code{backend = "plotly"}.
}
\details{
Note: Curves can be selected for plotting by passing a shorter \code{line_colors} vector.
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
