% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.multimarginal}
\alias{plot.multimarginal}
\title{Plots "multimarginal" Object}
\usage{
\method{plot}{multimarginal}(
  x,
  ncols = 2L,
  share_y = FALSE,
  sort = FALSE,
  ylim = NULL,
  scale_exposure = 1,
  line_colors = getOption("marginalplot.line_colors"),
  fill = getOption("marginalplot.fill"),
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)
}
\arguments{
\item{x}{An object of class "marginal".}

\item{ncols}{Number of columns in the plot layout.}

\item{share_y}{Should y axis be shared across all subplots?
No effect if \code{ylim} is passed.}

\item{sort}{Should plots be sorted in decreasing order of main effect importance
measured by the exposure weighted variance of partial dependence? Only if
partial dependence values have been calculated (i.e., not available for
\code{\link[=average_observed]{average_observed()}}).}

\item{ylim}{Manual y axis range.}

\item{scale_exposure}{Scale of the exposure bars, by default 1. Set to 0 for no bars.
Currently, values between 0 and 1 are possible only with backend "ggplot2".}

\item{line_colors}{Named vector of line colors. By default, a color blind
palette from "ggthemes" is used, equaling to
\code{c(obs = "#E69F00", pred = "#009E73", pd = "#56B4E9")}.
To change globally, set \code{options(marginalplot.line_colors = "named color vector")}.
Can be used to remove certain lines in the plot.}

\item{fill}{Fill color of bars. The default equals "lightgrey".
To change globally, set \code{options(marginalplot.fill = "new color")}.}

\item{wrap_x}{Should categorical x axis labels be wrapped after
this number of characters? The default is 10. Only for "ggplot2" backend.}

\item{rotate_x}{Should categorical x axis labels be rotated by this
number of degrees? Only for "ggplot2" backend. The default is 0 (no rotation).}

\item{backend}{Plotting backend, either "ggplot2" (default) or "plotly".
To change globally, set \code{options(marginalplot.backend = "plotly")}.}

\item{...}{Currently not used.}
}
\value{
Object of class "ggplot" (single plot) or "patchwork" (multiple plots),
or "plotly"/"htmlwidget" (with plotly backend).
}
\description{
Plots "multimarginal" Object
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
