% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Postprocess Output}
\usage{
postprocess(
  object,
  choose_stats = c("obs", "pred", "pd"),
  eval_at_center = FALSE,
  drop_below_n = 0,
  drop_below_prop = 0,
  explicit_na = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{object}{Object of class "marginal".}

\item{choose_stats}{Statistics to select, by default \code{c("obs", "pred", "pd")}.}

\item{eval_at_center}{If \code{FALSE} (default), the points are aligned with (weighted)
average X values per bin. If \code{TRUE}, the points are aligned with bar centers.
Since categoricals are always evaluated at bar centers, this only affects numerics.}

\item{drop_below_n}{Drop categories with exposure below this value.
Only for categorical X variables.}

\item{drop_below_prop}{Drop categories with relative exposure below this value.
Only for categorical X variables.}

\item{explicit_na}{Should \code{NA} levels be converted to strings?
Only for categorical X variables.}

\item{na.rm}{Should \code{NA} levels in X be dropped?}
}
\description{
This function offers different ways to improve the output of \code{\link[=marginal]{marginal()}},
\code{\link[=partial_dependence]{partial_dependence()}}, \code{\link[=average_observed]{average_observed()}}. Except for \code{choose_stats},
all arguments are vectorized, i.e., you can select different values per X variable.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5) |>
  postprocess(choose_stats = "pd", eval_at_center = TRUE) |>
  plot(num_points = TRUE)
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
