% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/average_observed.R
\name{average_observed}
\alias{average_observed}
\title{Average Observed}
\usage{
average_observed(
  x_name,
  data,
  y,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  ...
)
}
\arguments{
\item{x_name}{Column name of the stratification variable shown on the x axis.}

\item{data}{Matrix-like.}

\item{y}{Numeric vector with observed values of the response.
Can also be a column name in \code{data}. Omitted if \code{NULL} (default).}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{x_name}.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{x_name}.}

\item{discrete_m}{Numeric variable with up to this number of unique values
should not be binned. The default is 2. Set to \code{Inf} to avoid any binning.
Vectorized over \code{x_name}.}

\item{...}{Currently unused.}
}
\description{
Calculates average observed \code{y} values over one or multiple x variables specified
via \code{x_names}. \code{y} can either be a numeric vector or a column name in \code{data}.
This function is a convenience wrapper over \code{\link[=marginal]{marginal()}}.
}
\examples{
M <- average_observed(x_name = "Species", data = iris, y = "Sepal.Length")
M
M |> plot()

xvars <- c("Sepal.Width", "Species")
M <- average_observed(xvars, data = iris, y = "Sepal.Length")
M$Species
plot(M, rotate_x = 45)
}
