% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_observed.R
\name{average_observed}
\alias{average_observed}
\title{Average Observed}
\usage{
average_observed(
  v,
  data,
  y,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  ...
)
}
\arguments{
\item{v}{Vector of variable names to calculate statistics.}

\item{data}{Matrix or data.frame.}

\item{y}{Numeric vector with observed values of the response.
Can also be a column name in \code{data}. Omitted if \code{NULL} (default).}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, vector, string or function specifying the bins
of the numeric X variables as in \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
To allow varying values of \code{breaks} across variables, it can be a list of the
same length as \code{v}, or a \emph{named} list with \code{breaks} for certain variables.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{v}. Only relevant for numeric X.}

\item{discrete_m}{Numeric X variables with up to this number of unique values
should not be binned. The default is 2. Vectorized over \code{v}.}

\item{wprob_low}{Small values of numeric X variables are capped at this quantile.
Set to 0 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantiles (uses random seed). Vectorized over \code{v}.}

\item{wprob_high}{High values of numeric X variables are capped at this quantile.
Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (uses random seed). Vectorized over \code{v}.}

\item{...}{Currently unused.}
}
\value{
A list (of class "marginal") with a data.frame of statistics per feature. Use
single bracket subsetting to select part of the output.
}
\description{
Calculates average observed \code{y} values over one or multiple x variables specified
via \code{v}. \code{y} can either be a numeric vector or a column name in \code{data}.
This function is a convenience wrapper over \code{\link[=marginal]{marginal()}}.
}
\examples{
M <- average_observed(v = "Species", y = "Sepal.Length", data = iris)
M
M |> plot()
}
\seealso{
\code{\link[=marginal]{marginal()}}
}
