% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_observed.R
\name{average_observed}
\alias{average_observed}
\title{Average Observed}
\usage{
average_observed(
  X,
  y,
  w = NULL,
  x_name = "x",
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 13L,
  outlier_iqr = 2,
  seed = NULL,
  ...
)
}
\arguments{
\item{X}{A vector, matrix, or data.frame with features.}

\item{y}{A numeric vector representing observed response values.}

\item{w}{An optional numeric vector of weights. Having observations with
non-positive weight is equivalent to excluding them.}

\item{x_name}{If \code{X} is a vector: what is the name of the variable? By default "x".}

\item{breaks}{An integer, vector, or "Sturges" (the default) used to determine
bin breaks of continuous features. Values outside the total bin range are placed
in the outmost bins. To allow varying values of \code{breaks} across features,
\code{breaks} can be a list of the same length as \code{v}, or a \emph{named} list with breaks
for certain variables.}

\item{right}{Should bins be right-closed? The default is \code{TRUE}.
Vectorized over \code{v}. Only relevant for continuous features.}

\item{discrete_m}{Numeric features with up to this number of unique values should not
be binned but rather treated as discrete. The default is 13. Vectorized over \code{v}.}

\item{outlier_iqr}{If \code{breaks} is an integer or "Sturges", the breaks of a continuous
feature are calculated without taking into account feature values outside
quartiles +- \code{outlier_iqr} * IQR (where <= 9997 values are used to calculate the
quartiles). To let the breaks cover the full data range, set \code{outlier_iqr} to
0 or \code{Inf}. Vectorized over \code{v}.}

\item{seed}{Optional integer random seed used for calculating breaks:
The bin range is determined without values outside quartiles +- 2 IQR
using a sample of <= 9997 observations to calculate quartiles.}

\item{...}{Currently unused.}
}
\value{
A list (of class "EffectData") with a data.frame of statistics per feature. Use
single bracket subsetting to select part of the output.
}
\description{
Calculates average observed response over the values of one or multiple
variables specified by \code{X}. This describes the statistical association between the
response \code{y} and potential model features.
}
\details{
The function is a convenience wrapper around \code{\link[=feature_effects]{feature_effects()}}.
}
\examples{
M <- average_observed(iris$Species, y = iris$Sepal.Length)
M
M |> plot()

# Or multiple potential features X
average_observed(iris[2:5], y = iris[, 1], breaks = 5) |>
  plot()
}
\seealso{
\code{\link[=feature_effects]{feature_effects()}}
}
