% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Bias / Average Residuals}
\usage{
bias(
  X,
  resid,
  w = NULL,
  x_name = "x",
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 5L,
  outlier_iqr = 2,
  ...
)
}
\arguments{
\item{X}{A vector, matrix, or data.frame with variable(s) to be shown on the x axis.}

\item{resid}{A numeric vector of residuals, i.e., y - pred.}

\item{w}{An optional numeric vector of weights.}

\item{x_name}{If \code{X} is a vector: what is the name of the variable? By default "x".}

\item{breaks}{An integer, vector, string or function specifying the bins
of the numeric X variables as in \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
To allow varying values of \code{breaks} across variables, it can be a list of the
same length as \code{v}, or a \emph{named} list with \code{breaks} for certain variables.}

\item{right}{Should bins be right-closed? The default is \code{TRUE}.
Vectorized over \code{v}. Only relevant for numeric X.}

\item{discrete_m}{Numeric X variables with up to this number of unique values
should not be binned and treated as a factor (after calculating partial dependence)
The default is 5. Vectorized over \code{v}.}

\item{outlier_iqr}{Outliers of a numeric X are capped via the boxplot rule, i.e.,
outside \code{outlier_iqr} * IQR from the quartiles. The default is 2 is more
conservative than the usual rule to account for right-skewed distributions.
Set to 0 or \code{Inf} for no capping. Note that at most 10k observations are sampled
to calculate quartiles (uses random seed). Vectorized over \code{v}.}

\item{...}{Currently unused.}
}
\value{
A list (of class "marginal") with a data.frame of statistics per feature. Use
single bracket subsetting to select part of the output.
}
\description{
Calculates average residuals (= bias) over the values of one or multiple
\code{X} variables.
}
\details{
The function is a convenience wrapper around \code{\link[=marginal]{marginal()}}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
M <- bias(iris[2:5], resid = fit$residuals, breaks = 5)
M |> update(sort_by = "resid_mean") |> plot(share_y = "all")
}
\seealso{
\code{\link[=marginal]{marginal()}}
}
