% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{main_effect_importance}
\alias{main_effect_importance}
\title{Main Effect Importance}
\usage{
main_effect_importance(x, statistic = NULL)
}
\arguments{
\item{x}{Marginal object.}

\item{statistic}{The statistic used to calculate the variance for.
One of 'pd', 'pred', or 'obs'.}
}
\description{
Extracts the exposure weighted variances of the most relevant statistic
(pd > pred > obs) from a "marginal" object.
}
\seealso{
\code{\link[=postprocess]{postprocess()}}
}
