% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{main_effect_importance}
\alias{main_effect_importance}
\title{Main Effect Importance}
\usage{
main_effect_importance(x, statistic = NULL)
}
\arguments{
\item{x}{Marginal object.}

\item{statistic}{The statistic used to calculate the variance for.
One of 'pd', 'pred_mean', or 'y_mean'.}
}
\description{
Extracts the weighted variances of the most relevant statistic
(pd > pred_mean > y_mean) from a "marginal" object.
Serves as a simple "main effect" importance measure.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
M <- marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5)
main_effect_importance(M)
}
\seealso{
\code{\link[=postprocess]{postprocess()}}
}
