% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\alias{marginal.default}
\alias{marginal.ranger}
\alias{marginal.explainer}
\title{Marginal Statistics}
\usage{
marginal(object, ...)

\method{marginal}{default}(
  object,
  v,
  data,
  y = NULL,
  pred = NULL,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

\method{marginal}{ranger}(
  object,
  v,
  data,
  y = NULL,
  pred = NULL,
  pred_fun = NULL,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

\method{marginal}{explainer}(
  object,
  v,
  data = object[["data"]],
  y = NULL,
  pred = NULL,
  pred_fun = object[["predict_function"]],
  w = object[["weights"]],
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)
}
\arguments{
\item{object}{Fitted model.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} or (typically) \code{prob = TRUE} in binary probabilistic models.}

\item{v}{Vector of variable names to be used on the x axis.}

\item{data}{Matrix-like.}

\item{y}{Numeric vector with observed values of the response.
Can also be a column name in \code{data}. Omitted if \code{NULL} (default).}

\item{pred}{Numeric vector with predictions. If \code{NULL}, it is calculated as
\code{pred_fun(object, data, ...)}. Used to save time if \code{marginal()} is to be
called multiple times.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{v}. Only relevant for numeric x.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{v}. Only relevant for numeric x.}

\item{discrete_m}{Numeric x with up to this number of unique values
should be treated as factors. The default is 2. Vectorized over \code{v}.}

\item{wprob_low}{Small values of numeric x are capped at this quantile.
Set to 0 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{v}.}

\item{wprob_high}{High values of numeric x are capped at this quantile.
Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{v}.}

\item{calc_pred}{Should predictions be calculated? Default is \code{TRUE}. Only relevant
if \code{pred = NULL}.}

\item{pd_n}{Size of the data used for calculation of partial dependence.
The default is 500. Set to 0 (or pass \code{pred_fun = NULL}) to omit calculation
of partial dependence. This depends on your random seed.}
}
\value{
If \code{v} has length 1, an object of class "marginal" containing these elements:
\itemize{
\item \code{data}: data.frame containing statistics and plot positions of values and bars.
\item \code{num}: Indicator whether x is numeric.
\item \code{v}: Same as input \code{v}.
If \code{v} has length > 1, an object of class "multimarginal", which is a named
list of "marginal" objects.
}
}
\description{
This is the main function of the package. It calculates
\itemize{
\item average observed,
\item average predicted,
\item partial dependence, and
\item counts/weights
over a (possibly binned) feature v, optionally weighted with weights \code{w}.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{marginal(default)}: Default method.

\item \code{marginal(ranger)}: Method for "ranger" models.

\item \code{marginal(explainer)}: Method for DALEX "explainer".

}}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
M <- marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5)
M
M |> plot()
}
\seealso{
\code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
