% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\alias{partial_dependence.default}
\alias{partial_dependence.ranger}
\alias{partial_dependence.explainer}
\title{Partial Dependence}
\usage{
partial_dependence(object, ...)

\method{partial_dependence}{default}(
  object,
  x_name,
  data,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{ranger}(
  object,
  x_name,
  data,
  pred_fun = NULL,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{explainer}(
  object,
  x_name,
  data = object[["data"]],
  pred_fun = object[["predict_function"]],
  w = object[["weights"]],
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  winsorize_low = 0.01,
  winsorize_high = 0.99,
  pd_n = 500L,
  ...
)
}
\arguments{
\item{object}{Fitted model.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} or (typically) \code{prob = TRUE} in binary probabilistic models.}

\item{x_name}{Column name of the variable shown on the x axis.}

\item{data}{Matrix-like.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, a vector, a string or a function specifying the bins
of \code{x}, and passed to \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
\emph{Not} vectorized over \code{x_name}. Only relevant for numeric x.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{x_name}. Only relevant for numeric x.}

\item{discrete_m}{Numeric x with up to this number of unique values
should be treated as factors. The default is 2. Vectorized over \code{x_name}.}

\item{winsorize_low}{Small values of numeric x are capped at this quantile.
Set to 0 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{x_name}.}

\item{winsorize_high}{High values of numeric x are capped at this quantile.
Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled
to calculate the quantile (depends on your random seed). Vectorized over \code{x_name}.}

\item{pd_n}{Size of the data used for calculation of partial dependence.
The default is 500. Set to 0 (or pass \code{pred_fun = NULL}) to omit calculation
of partial dependence. This depends on your random seed.}
}
\value{
If \code{x_name} has length 1, an object of class "marginal" containing these elements:
\itemize{
\item \code{data}: data.frame containing statistics and plot positions of values and bars.
\item \code{num}: Indicator whether x is numeric.
\item \code{x_name}: Same as input \code{x_name}.
If \code{x_name} has length > 1, an object of class "multimarginal", which is a named
list of "marginal" objects.
}
}
\description{
Calculates partial dependence of one or more variables (\code{x_name}) in
a dataset \code{data}. This function is a convenience wrapper over \code{\link[=marginal]{marginal()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{partial_dependence(default)}: Default method.

\item \code{partial_dependence(ranger)}: Default method.

\item \code{partial_dependence(explainer)}: Default method.

}}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
M <- partial_dependence(fit, x_name = "Species", data = iris)
M
M |> plot()
}
\references{
Friedman, Jerome H. \emph{"Greedy Function Approximation: A Gradient Boosting Machine."}
Annals of Statistics 29, no. 5 (2001): 1189-1232.
}
\seealso{
\code{\link[=marginal]{marginal()}}
}
