% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\alias{partial_dependence.default}
\alias{partial_dependence.ranger}
\alias{partial_dependence.explainer}
\title{Partial Dependence}
\usage{
partial_dependence(object, ...)

\method{partial_dependence}{default}(
  object,
  v,
  data,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{ranger}(
  object,
  v,
  data,
  pred_fun = NULL,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  pd_n = 500L,
  ...
)

\method{partial_dependence}{explainer}(
  object,
  v,
  data = object[["data"]],
  pred_fun = object[["predict_function"]],
  w = object[["weights"]],
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  pd_n = 500L,
  ...
)
}
\arguments{
\item{object}{Fitted model.}

\item{...}{Further arguments passed to \code{pred_fun()}, e.g., \code{type = "response"} in
a \code{glm()} or (typically) \code{prob = TRUE} in binary probabilistic models.}

\item{v}{Vector of variable names to calculate statistics.}

\item{data}{Matrix or data.frame.}

\item{pred_fun}{Prediction function, by default \code{stats::predict}.
The function takes three arguments (names irrelevant): \code{object}, \code{data}, and \code{...}.}

\item{w}{Optional vector with case weights. Can also be a column name in \code{data}.}

\item{breaks}{An integer, vector, string or function specifying the bins
of the numeric X variables as in \code{\link[graphics:hist]{graphics::hist()}}. The default is "Sturges".
To allow varying values of \code{breaks} across variables, it can be a list of the
same length as \code{v}, or a \emph{named} list with \code{breaks} for certain variables.}

\item{right}{Should bins created via \code{\link[graphics:hist]{graphics::hist()}} be right-closed?
The default is \code{TRUE}. Vectorized over \code{v}. Only relevant for numeric X.}

\item{discrete_m}{Numeric X variables with up to this number of unique values
should not be binned. The default is 2. Vectorized over \code{v}.}

\item{wprob_low}{The smallest values of a numeric X are capped at the corresponding
quantile. Set to 0 to avoid Winsorizing. Note that at most 100k observations
are sampled to calculate the quantile (uses random seed). Vectorized over \code{v}.}

\item{wprob_high}{The largest values of a numeric X are capped at the corresponding
quantile. Set to 1 to avoid Winsorizing. Note that at most 100k observations
are sampled to calculate the quantile (uses random seed). Vectorized over \code{v}.}

\item{pd_n}{Size of the data used for calculating partial dependence.
The default is 500. For larger \code{data} (and \code{w}), \code{pd_n} rows are randomly sampled
(uses random seed). Each variable specified by \code{v} uses the same subsample.
Set to 0 to omit.}
}
\value{
A list (of class "marginal") with a data.frame of statistics per feature. Use
single bracket subsetting to select part of the output.
}
\description{
Calculates partial dependence of one or more X variables (\code{v}) in \code{data}.
This function is a convenience wrapper around \code{\link[=marginal]{marginal()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{partial_dependence(default)}: Default method.

\item \code{partial_dependence(ranger)}: Default method.

\item \code{partial_dependence(explainer)}: Default method.

}}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
M <- partial_dependence(fit, v = "Species", data = iris)
M |> plot()

M2 <- partial_dependence(fit, v = colnames(iris)[-1], data = iris)
plot(M2, share_y = TRUE)
}
\references{
Friedman, Jerome H. \emph{"Greedy Function Approximation: A Gradient Boosting Machine."}
Annals of Statistics 29, no. 5 (2001): 1189-1232.
}
\seealso{
\code{\link[=marginal]{marginal()}}
}
