% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marginal}
\alias{plot.marginal}
\title{Plots Marginal Object}
\usage{
\method{plot}{marginal}(
  x,
  ncols = 2L,
  share_y = FALSE,
  ylim = NULL,
  scale_bars = 1,
  cat_lines = TRUE,
  num_points = FALSE,
  colors = getOption("marginalplot.colors"),
  fill = getOption("marginalplot.fill"),
  bar_width = 0.7,
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)
}
\arguments{
\item{x}{An object of class "marginal".}

\item{ncols}{Number of columns in the plot layout.
Only if \code{length(x) > 1} (multiple plots). With "plotly" subplots, the result may
differ slightly.}

\item{share_y}{Should y axis be shared across all subplots?
No effect if \code{ylim} is passed. Only if \code{length(x) > 1} (multiple plots).}

\item{ylim}{Manual y axis range.}

\item{scale_bars}{Vertical scaling of the bars (between 0 and 1).
The default is 1. Set to 0 for no bars.}

\item{cat_lines}{Show lines for non-numeric features. Default is \code{TRUE}.
Vectorized over \code{x}.}

\item{num_points}{Show points for numeric features. Default is \code{FALSE}.
Vectorized over \code{x}.}

\item{colors}{Line colors in the order "obs", "pred", "pd" (or a subset thereof).
By default, a color blind friendly palette from "ggthemes", namely
\code{c("#CC79A7", "#009E73", "#56B4E9")}.
To change globally, set \verb{options(marginalplot.colors = new colors)}.}

\item{fill}{Fill color of bars. The default equals "lightgrey".
To change globally, set \verb{options(marginalplot.fill = new color)}.}

\item{bar_width}{Relative bar width of non-numeric features, by default 0.7.}

\item{wrap_x}{Should categorical xaxis labels be wrapped after this length?
The default is 10. Set to 0 for no wrapping. Vectorized over \code{x}.
Only for "ggplot2" backend.}

\item{rotate_x}{Should categorical xaxis labels be rotated by this angle?
The default is 0 (no rotation). Vectorized over \code{x}. Only for "ggplot2" backend.}

\item{backend}{Plot backend, either "ggplot2" (default) or "plotly".
To change globally, set \code{options(marginalplot.backend = "plotly")}.}

\item{...}{Currently unused.}
}
\value{
If \code{length(x) == 1} (single plot), an object of class  "ggplot" or "plotly".
Otherwise, an object of class "patchwork", or a "plotly" subplot.
}
\description{
Plots marginal statistics using a color blind palette from "ggthemes".
You can switch to plotly by setting \code{backend = "plotly"}.
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
