% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marginal}
\alias{plot.marginal}
\title{Plots Marginal Object}
\usage{
\method{plot}{marginal}(
  x,
  ncol = 2L,
  byrow = TRUE,
  share_y = FALSE,
  ylim = NULL,
  cat_lines = TRUE,
  num_points = FALSE,
  title = "",
  subplot_titles = TRUE,
  ylab = NULL,
  lines = c(obs = "y_mean", pred = "pred_mean", pd = "pd"),
  colors = getOption("marginalplot.colors"),
  fill = getOption("marginalplot.fill"),
  bar_height = 1,
  bar_width = 0.7,
  bar_measure = c("weight", "N"),
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)
}
\arguments{
\item{x}{An object of class "marginal".}

\item{ncol}{Number of columns in the plot layout. Only for multiple plots.}

\item{byrow}{Should plots be placed by row? Default is \code{TRUE}.
Only for multiple plots.}

\item{share_y}{Should y axis be shared across all subplots?
No effect if \code{ylim} is passed. Only for multiple plots.}

\item{ylim}{A vector of length 2 with manual y axis limits.}

\item{cat_lines}{Show lines for non-numeric features. Default is \code{TRUE}.}

\item{num_points}{Show points for numeric features. Default is \code{FALSE}.}

\item{title}{Overall plot title, by default \code{""} (no title).}

\item{subplot_titles}{Should variable names be shown as subplot titles?
Default is \code{TRUE}. Only for multiple plots.}

\item{ylab}{Label of the y axis. The default \code{NULL} automatically derives
a reasonable name based on the calculated statistics.}

\item{lines}{Named vector of the form \code{c(name = statistic, ...)} of statistics to
plot. \code{name} is used as legend label. By default
\code{c(obs = "y_mean", pred = "pred_mean", pd = "pd")}.
To show only bias with "bias" as legend entry, use \code{c(bias = "resid_mean")}.}

\item{colors}{Vector of line colors of sufficient length.
By default, a color blind friendly palette from "ggthemes".
To change globally, set \verb{options(marginalplot.colors = new colors)}.}

\item{fill}{Fill color of bars. The default equals "lightgrey".
To change globally, set \verb{options(marginalplot.fill = new color)}.}

\item{bar_height}{Relative bar height (default 1). Set to 0 for no bars.}

\item{bar_width}{Relative bar width of non-numeric features, by default 0.7.}

\item{bar_measure}{What should bars represent? Either "weight" (default) or "N".}

\item{wrap_x}{Should categorical x axis labels be wrapped after this length?
The default is 10. Set to 0 for no wrapping. Vectorized over \code{x}.
Only for "ggplot2" backend.}

\item{rotate_x}{Should categorical xaxis labels be rotated by this angle?
The default is 0 (no rotation). Vectorized over \code{x}. Only for "ggplot2" backend.}

\item{backend}{Plot backend, either "ggplot2" (default) or "plotly".
To change globally, set \code{options(marginalplot.backend = "plotly")}.}

\item{...}{Passed to \code{patchwork::plot_layout()} or \code{plotly::subplot()}. Typically
not used.}
}
\value{
If single plot, an object of class  "ggplot" or "plotly".
Otherwise, an object of class "patchwork", or a "plotly" subplot.
}
\description{
Plots marginal statistics using a color blind palette from "ggthemes".
You can switch to plotly by setting \code{backend = "plotly"}.
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
