% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marginal}
\alias{plot.marginal}
\title{Plots "marginal" Object}
\usage{
\method{plot}{marginal}(
  x,
  statistics = c("y_mean", "pred_mean", "pd", "ale"),
  ncol = grDevices::n2mfrow(length(x))[2L],
  byrow = TRUE,
  share_y = c("no", "all", "rows", "cols"),
  ylim = NULL,
  cat_lines = TRUE,
  num_points = FALSE,
  title = "",
  subplot_titles = TRUE,
  ylab = NULL,
  legend_labels = NULL,
  colors = getOption("effectplots.colors"),
  fill = getOption("effectplots.fill"),
  alpha = 1,
  bar_height = 1,
  bar_width = 0.7,
  bar_measure = c("weight", "N"),
  wrap_x = 10,
  rotate_x = 0,
  plotly = getOption("effectplots.plotly"),
  ...
)
}
\arguments{
\item{x}{An object of class "marginal".}

\item{statistics}{Vector of (available) statistics to show. By default
\code{c("y_mean", "pred_mean", "pd", "ale")}. E.g., used to hide certain
statistics, or to show only \code{"resid_mean"}. Note that the \code{statistics} argument
is automatically set to \code{"resid_mean"} if the result of \code{\link[=bias]{bias()}} is to be plotted.
Additionally, it controls the order in which lines are added to the plot.}

\item{ncol}{Number of columns of the plot layout, by default
\code{grDevices::n2mfrow(length(x))[2L]}. Only relevant for multiple plots.}

\item{byrow}{Should plots be placed by row? Default is \code{TRUE}.
Only for multiple plots.}

\item{share_y}{Should y axis be shared across subplots? The default is "no". Other
choices are "all", "rows", and "cols".
No effect if \code{ylim} is passed. Only for multiple plots.}

\item{ylim}{A vector of length 2 with manual y axis limits, or a list thereof.}

\item{cat_lines}{Show lines for non-numeric features. Default is \code{TRUE}.}

\item{num_points}{Show points for numeric features. Default is \code{FALSE}.}

\item{title}{Overall plot title, by default \code{""} (no title).}

\item{subplot_titles}{Should variable names be shown as subplot titles?
Default is \code{TRUE}. Only for multiple plots.}

\item{ylab}{Label of the y axis. The default \code{NULL} automatically derives
a reasonable name.}

\item{legend_labels}{Vector of legend labels corresponding to \code{statistics}.
Either \code{NULL} or a vector of the same length as \code{statistics}.}

\item{colors}{Vector of line/point colors at least as long as "statistics".
By default, a color blind friendly palette from "ggthemes" with five values.
To change globally, set \verb{options(effectplots.colors = new colors)}.}

\item{fill}{Fill color of bars. The default equals "lightgrey".
To change globally, set \verb{options(effectplots.fill = new color)}.}

\item{alpha}{Alpha transparency of lines and points. Default is 1.}

\item{bar_height}{Relative bar height (default 1). Set to 0 for no bars.}

\item{bar_width}{Relative bar width of non-numeric features, by default 0.7.}

\item{bar_measure}{What should bars represent? Either "weight" (default) or "N".}

\item{wrap_x}{Should categorical x axis labels be wrapped after this length?
The default is 10. Set to 0 for no wrapping. Vectorized over \code{x}.
Only for "ggplot2" backend.}

\item{rotate_x}{Should categorical xaxis labels be rotated by this angle?
The default is 0 (no rotation). Vectorized over \code{x}. Only for "ggplot2" backend.}

\item{plotly}{Should 'plotly' be used? The default is \code{FALSE} ('ggplot2' with
'patchwork'). Use \code{options(effectplots.plotly = TRUE)} to change globally.}

\item{...}{Passed to \code{patchwork::plot_layout()} or \code{plotly::subplot()}. Typically
not used.}
}
\value{
If a single plot, an object of class  "ggplot" or "plotly".
Otherwise, an object of class "patchwork", or a "plotly" subplot.
}
\description{
Versatile plot function for a "marginal" object. Set \code{plotly = TRUE} for
interactive plots.
By default, a color blind palette from "ggthemes" is used.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
M <- marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5)
plot(M, share_y = "all")
plot(M, statistics = c("pd", "ale"), legend_labels = c("PD", "ALE"))
plot(M, statistics = "resid_mean", share_y = "all")
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=average_predicted]{average_predicted()}},
\code{\link[=partial_dependence]{partial_dependence()}}, \code{\link[=bias]{bias()}}, \code{\link[=ale]{ale()}}
}
