% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\alias{postprocess.default}
\alias{postprocess.marginal}
\alias{postprocess.multimarginal}
\title{Postprocess Output of marginal()}
\usage{
postprocess(object, ...)

\method{postprocess}{default}(object, ...)

\method{postprocess}{marginal}(
  object,
  cat_drop_below_n = 0,
  cat_drop_below_prop = 0,
  cat_explicit_na = TRUE,
  drop_na = FALSE,
  ...
)

\method{postprocess}{multimarginal}(object, ...)
}
\arguments{
\item{object}{Object of class "marginal" or "multimarginal".}

\item{...}{Currently unused.}

\item{cat_drop_below_n}{Drop categories with exposure below this value.}

\item{cat_drop_below_prop}{Drop categories with relative exposure below this value.}

\item{cat_explicit_na}{Should \code{NA} levels in categoricals be converted to strings?}

\item{drop_na}{Should \code{NA} levels be dropped?}
}
\description{
This function offers different ways to improve the output of \code{\link[=marginal]{marginal()}},
\code{\link[=partial_dependence]{partial_dependence()}}, \code{\link[=average_observed]{average_observed()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{postprocess(default)}: Default method.

\item \code{postprocess(marginal)}: Method for "marginal" object.

\item \code{postprocess(multimarginal)}: Method for "multimarginal" object.

}}
