% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Postprocess Marginal Object}
\usage{
postprocess(
  x,
  sort = FALSE,
  sort_by = NULL,
  collapse_m = 30L,
  collapse_by = c("weight", "N"),
  drop_below_n = 0,
  drop_below_weight = 0,
  na.rm = FALSE
)
}
\arguments{
\item{x}{Object of class "marginal".}

\item{sort}{Should \code{x} be sorted in decreasing order of feature importance?
Importance is measured by the weighted variance of the most relevant available
statistic (pd > pred_mean > y_mean > resid_mean). The default is \code{FALSE}.}

\item{sort_by}{If \code{sort = TRUE}, this measure used for sorting. One of
'pd', 'pred_mean', 'y_mean', or 'resid_mean' (if available). The default is \code{NULL},
which picks the first of the listed statistics.}

\item{collapse_m}{If a categorical X has more than \code{collapse_m} levels,
rare levels are collapsed into a new level "Other". Standard deviations are
collapsed via root of the weighted average variances. By default 30.
Set to \code{Inf} for no collapsing. Vectorized over \code{x}.}

\item{collapse_by}{How to determine "rare" levels in \code{collapse_m}?
Either "weight" (default) or "N". Only matters in situations with case weights \code{w}.}

\item{drop_below_n}{Drop bins with N below this value. Applied after collapsing.
Vectorized over \code{x}.}

\item{drop_below_weight}{Drop bins with weight below this value. Applied after
collapsing. Vectorized over \code{x}.}

\item{na.rm}{Should missing bin centers be dropped? Default is \code{FALSE}.}
}
\description{
This function helps to improve the output of \code{\link[=marginal]{marginal()}},
\code{\link[=partial_dependence]{partial_dependence()}}, and \code{\link[=average_observed]{average_observed()}}.
}
\details{
If \code{sort = TRUE}, the features will be sorted by a simple variable importance
measure. It is calculated as the variance of the most relevant available statistic
(pd > pred_mean > y_mean). The variance is weighted with bin weights to reflect
the data distribution. For "pd", this measure is similar (but not identical) to the
suggestion of Greenwell et al. (2018).
Importance is calculated after the other postprocessings, e.g., after collapsing
rare levels.

Except for \code{sort}, all arguments are vectorized, i.e., you can
pass a vector or list of the same length as \code{x}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5) |>
  postprocess(sort = TRUE) |>
  plot()
}
\references{
Greenwell, Brandon M., Bradley C. Boehmke, and Andrew J. McCarthy.
\emph{A Simple and Effective Model-Based Variable Importance Measure.} Arxiv (2018).
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}},
\code{\link[=main_effect_importance]{main_effect_importance()}}
}
