% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Postprocess Marginal Object}
\usage{
postprocess(
  x,
  sort = FALSE,
  drop_stats = NULL,
  eval_at_center = FALSE,
  collapse_m = 30L,
  drop_below_n = 0,
  na.rm = FALSE
)
}
\arguments{
\item{x}{Object of class "marginal".}

\item{sort}{Should \code{x} be sorted in decreasing order of feature importance?
Importance is measured by the weighted variance of the most relevant available
statistic (pd > pred > obs), see Details. The default is \code{FALSE}.}

\item{drop_stats}{Statistics to drop, by default \code{NULL}.
Subset of "pred", "obs", "pd". Not vectorized over \code{x}.}

\item{eval_at_center}{If \code{FALSE} (default), the points are aligned with (weighted)
average X values per bin. If \code{TRUE}, the points are aligned with bar centers.
Since categorical X are always evaluated at bar centers, this only affects
numeric X. Note that partial dependence of numeric X is always evaluated at
bar means, not centers. Vectorized over \code{x}.}

\item{collapse_m}{If a categorical X has more than \code{collapse_m} levels,
rare levels are collapsed into a new level "Other". Standard deviations are
collapsed via root of the weighted average variances.
By default 30. Set to \code{Inf} for no collapsing. Vectorized over \code{x}.}

\item{drop_below_n}{Drop bins with weight below this value. Applied after the
effect of \code{collapse_m}. Vectorized over \code{x}.}

\item{na.rm}{Should \code{NA} levels in X be dropped?}
}
\description{
This function helps to improve the output of \code{\link[=marginal]{marginal()}},
\code{\link[=partial_dependence]{partial_dependence()}}, and \code{\link[=average_observed]{average_observed()}}.
}
\details{
If \code{sort = TRUE}, the features will be sorted by a simple variable importance
measure. It is calculated as the variance of the most relevant available statistic
(pd > pred > obs). The variance is weighted by N to reflect the data distribution.
For "pd", this measure is similar (but not identical) to the suggestion of
Greenwell et al. (2018).
Importance is calculated after the other postprocessings, e.g., after collapsing
rare levels.

Except for \code{sort} and \code{drop_stats}, all arguments are vectorized, i.e., you can
pass a vector or list of the same length as \code{x}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5) |>
  postprocess(sort = TRUE, eval_at_center = TRUE) |>
  plot(num_points = TRUE)
}
\references{
Greenwell, Brandon M., Bradley C. Boehmke, and Andrew J. McCarthy.
\emph{A Simple and Effective Model-Based Variable Importance Measure.} Arxiv (2018).
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
