% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Postprocess Marginal Object}
\usage{
postprocess(
  x,
  sort = FALSE,
  drop_stats = NULL,
  eval_at_center = FALSE,
  collapse_m = Inf,
  explicit_na = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{x}{Object of class "marginal".}

\item{sort}{Should \code{x} be sorted in decreasing order of feature importance?
Importance is measured by the exposure weighted variance of the most
relevant available statistic (pd > pred > obs). The default is \code{FALSE}.
Importance is calculated after the other postprocessing steps.}

\item{drop_stats}{Statistics to drop, by default \code{NULL}.
Subset of "pred", "obs", "pd". Not vectorized over \code{x} elements.}

\item{eval_at_center}{If \code{FALSE} (default), the points are aligned with (weighted)
average X values per bin. If \code{TRUE}, the points are aligned with bar centers.
Since categoricals are always evaluated at bar centers, this only affects numerics.
Note that partial dependence of numeric X is always evaluated at bar means, not
centers.}

\item{collapse_m}{If a categorical X has more than \code{collapse_m} levels,
low exposure levels are collapsed into a new level "Other".
By default \code{Inf} (no collapsing).}

\item{explicit_na}{Should \code{NA} levels be converted to strings?
Only for categorical X variables.}

\item{na.rm}{Should \code{NA} levels in X be dropped?}
}
\description{
This function helps to improve the output of \code{\link[=marginal]{marginal()}},
\code{\link[=partial_dependence]{partial_dependence()}}, \code{\link[=average_observed]{average_observed()}}. Except for \code{sort} and \code{drop_stats},
all arguments are vectorized, i.e., you can pass a list of the same length as
\code{x}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5) |>
  postprocess(sort = TRUE, eval_at_center = TRUE) |>
  plot(num_points = TRUE)
}
\seealso{
\code{\link[=marginal]{marginal()}}, \code{\link[=average_observed]{average_observed()}}, \code{\link[=partial_dependence]{partial_dependence()}}
}
